+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.6
+* Supported [vega](https://vega.github.io/vega/) and [vega-lite](https://vega.github.io/vega-lite/).   
+* Supported [ditaa](https://github.com/stathissideris/ditaa).  
+* Changed the key-value pair way of writing attributes to the pandoc way.   
+* Supported multiple new preview themes and code block themes ported from atom.    
+
 # 0.1.5
 * Fixed `webview.ts` `clickTagA` action bug.